chore(utilities): Rename sub-directory
