+++ packages/utilities/src/index.ts
@@ -4,4 +4,4 @@ export * from './css-utilities'
 export * from './dom'
 export * from './functional'
 export * from './string-utilities'
-export * from './typescript-utilities'
+export * from './types'
