+++ packages/utilities/src/types/Try.ts
